css-animations-1/pseudoElement-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS3 Animations Test:  pseudoElement attribute of Event - return the name or empty string</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Jieqiong Cui" href="mailto:jieqiongx.cui@intel.com">
<link rel="help" title="5.1.2. Attributes" href="http://www.w3.org/TR/css3-animations/#AnimationEvent-attributes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes sample {
    from {
      left: 150px;
    }
    to {
      left: 0px;
    }
  }
  .test {
    animation-duration: 10s;
    animation-name: sample;
    animation-timing-function: linear;
    background-color: blue;
    height: 50px;
    position: relative;
    width: 100px;
  }
  #testDiv2::after {
    content: "pseudoElement";
  }
</style>
<div id="log"></div>
<div class="test" id="testDiv1"></div>
<div class="test" id="testDiv2"></div>
<script>
  var t1 = async_test("Check if the pseudoElement attribute returns the empty string when the animation runs on an element");
  var t2 = async_test("Check if the pseudoElement attribute returns the name (beginning with two colons) of the CSS pseudo-element when the animation runs");
  var testDiv1 = document.getElementById('testDiv1');
  var testDiv2 = document.getElementById('testDiv2');

  testDiv1.addEventListener('animationstart', function(event) {
    t1.step(function() {
      assert_equals(event.pseudoElement, "", "the pseudoElement value");
    });
    t1.done();
  }, false);

  testDiv2.addEventListener('animationstart', function(event) {
    t2.step(function() {
      assert_equals(event.pseudoElement, "::after", "the pseudoElement value");
    });
    t2.done();
  }, false);
</script>
